+++ packages/dht/src/proto/DhtRpc.ts
@@ -474,9 +474,17 @@ export interface LockResponse {
     reason?: boolean; // if accepted = false
 }
 /**
- * @generated from protobuf message Confirmation
+ * @generated from protobuf message DisconnectNotice
  */
-export interface Confirmation {
+export interface DisconnectNotice {
+    /**
+     * @generated from protobuf field: string protocolVersion = 1;
+     */
+    protocolVersion: string;
+    /**
+     * @generated from protobuf field: PeerDescriptor peerDescriptor = 2;
+     */
+    peerDescriptor?: PeerDescriptor;
 }
 /**
  * @generated from protobuf enum NodeType
@@ -902,15 +910,18 @@ class LockResponse$Type extends MessageType$<LockResponse> {
  */
 export const LockResponse = new LockResponse$Type();
 // @generated message type with reflection information, may provide speed optimized methods
-class Confirmation$Type extends MessageType$<Confirmation> {
+class DisconnectNotice$Type extends MessageType$<DisconnectNotice> {
     constructor() {
-        super("Confirmation", []);
+        super("DisconnectNotice", [
+            { no: 1, name: "protocolVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "peerDescriptor", kind: "message", T: () => PeerDescriptor }
+        ]);
     }
 }
 /**
- * @generated MessageType for protobuf message Confirmation
+ * @generated MessageType for protobuf message DisconnectNotice
  */
-export const Confirmation = new Confirmation$Type();
+export const DisconnectNotice = new DisconnectNotice$Type();
 /**
  * @generated ServiceType for protobuf service DhtRpcService
  */
@@ -941,5 +952,6 @@ export const WebRtcConnectorService = new ServiceType("WebRtcConnectorService", 
  */
 export const ConnectionLocker = new ServiceType("ConnectionLocker", [
     { name: "lockRequest", options: {}, I: LockRequest, O: LockResponse },
-    { name: "unlockRequest", options: {}, I: UnlockRequest, O: Empty }
+    { name: "unlockRequest", options: {}, I: UnlockRequest, O: Empty },
+    { name: "gracefulDisconnect", options: {}, I: DisconnectNotice, O: Empty }
 ]);