+++ packages/network/src/composition.ts
@@ -11,8 +11,10 @@ export {
 } from './identifiers'
 export {
     COUNTER_UNSUBSCRIBE,
-    DEFAULT_MAX_NEIGHBOR_COUNT
+    DEFAULT_MAX_NEIGHBOR_COUNT,
+    STREAMR_ICE_SERVERS
 } from './constants'
+export { IceServer } from './connection/webrtc/WebRtcConnection'
 export { NetworkNode } from './logic/NetworkNode'
 export { Event as NodeEvent } from './logic/Node'
 export { NameDirectory } from './NameDirectory'