+++ packages/trackerless-network/bin/full-node-websocket.ts
@@ -31,7 +31,14 @@ async function run(): Promise<void> {
         websocket: { ip: program.opts().entrypointIp, port: 23123 }
     }
 
-    const layer0 = new DhtNode({ webSocketPort: port, webSocketHost: ip, peerIdString: program.opts().id, numberOfNodesPerKBucket: 2 })
+    const layer0 = new DhtNode({
+        webSocketPort: port,
+        webSocketHost: ip,
+        peerIdString: program.opts().id,
+        numberOfNodesPerKBucket: 2,
+        entryPoints: [epPeerDescriptor]
+    })
+
     await layer0.start()
 
     await layer0.joinDht(epPeerDescriptor)